Initial Commit
This commit consists of the first Draft of the

Third Person Styled game I have been working on for mobile & web

Currently, there is a Player set-up and a Sample Environment created by me to test simple mechanics for the game.

And an Enemy type of Drone which if near the player will attack him with rubber balls and bullets.